api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
boto3==1.34.162
python-multipart==0.0.9
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
kafka-python-ng==2.2.2
mlflow==2.15.1
six==1.17.0
